zap-doc: refactor CondenseZap
The old implementation finds all paragraph breaks, checks if the text on
both sides is highlighted, and if so replaces the paragraph break with a
space.

This causes issues when there are multiple paragraph breaks in a row,
which are preserved in the final document.

The new implementation goes through all Tags and uses a Verbatim
function to select the card under it, then swaps paragraph breaks with
spaces in that card.

This also runs faster, since it merely has to process all cards instead
of all paragraph breaks.

Fixes: #1 (CondenseZap Extra Line Breaks)