zapper: implement granular zapper
One requested feature was to be able to selectively zap certain portions
of the document. This changes the zap helper functions to take a range
as an argument. The zap doc macro passes the entire document as a range,
but you can also place your cursor in the tag of a card to pass the card
as the range.

Closes: https://github.com/shreerammodi/debate-scripts/issues/3