zapper: send-doc: ensure compatibility with Michigan Analytics style
Even though this style is discouraged to use, including compatibility
only minutely impacts performance. We first check if the style exists in
the document, and then perform a search for text with that style, which
means that documents that don't use the Analytics style will have the
same performance as before, but compatibility is ensured for documents
that use Analytics.