forreference: implement fast and slow versions
The fast version, which operates on runs of highlighted characters,
sometimes removes formatting such as emphasis. This is unintended, but
cannot be avoided with the fast implementation.

So, we provide a slow version, which deosn't run into these errors and
can be run when time isn't a concern, and a fast version which is much
more performant.